requirements.txt
kafka-python==2.0.2
minio==7.1.15
pyspark==3.3.0
python-dotenv==1.0.0
requests==2.31.0  
pandas==2.0.3     
pyarrow==14.0.1   #